magento/adobe-stock-integration#1504: [2.1-develop] Insert rendition images to the content from media gallery instead of original images - Refactor based on PR suggestions
