+++ MediaGalleryRenditions/composer.json
@@ -7,7 +7,6 @@
         "magento/module-media-gallery-renditions-api": "*",
         "magento/module-media-gallery-synchronization-api": "*",
         "magento/module-media-content-api": "*",
-        "magento/module-media-content": "*",
         "magento/module-cms": "*"
     },
     "type": "magento2-module",
