+++ MediaGalleryRenditions/etc/media_content.xml
@@ -1,0 +1,14 @@
+<?xml version="1.0"?>
+<!--
+/**
+ * Copyright © Magento, Inc. All rights reserved.
+ * See COPYING.txt for license details.
+ */
+-->
+<config xmlns:xsi="http://www.w3.org/2001/XMLSchema-instance" xsi:noNamespaceSchemaLocation="urn:magento:module:Magento_MediaContentApi:etc/media_content.xsd">
+    <search>
+        <patterns>
+            <pattern name="renditions">/"?.renditions(.*?)"?}}/</pattern>
+        </patterns>
+    </search>
+</config>
